distribute things
